宣言を抽象クラスにすることでコードをまとめた
