+++ src/commands/button/DeclarationSecond.ts
@@ -1,136 +1,17 @@
-import { ButtonBuilder, ButtonStyle, ButtonInteraction, Guild } from "discord.js";
-import dayjs from "dayjs";
+import { DeclarationAbstract } from '@/commands/button/DeclarationAbstract'
+import { ButtonBuilder, ButtonStyle } from 'discord.js';
 
-import Declaration from "@/app/model/Declaration";
-import DataSource from "@/datasource";
-import Boss from "@/entity/Boss";
-import Clan from "@/entity/Clan";
-import Lap from "@/entity/Lap";
-import Event from "@/entity/Event";
-import DeclarationRepository from "@/entity/Declaration";
-import BossChannelMessage from "@/messages/BossChannelMessage";
-import EventBoss from "@/entity/EventBoss";
+export class DeclarationSecond extends DeclarationAbstract {
+  static readonly customId: string = "declarationSecond";
+  attackCount: number;
+  data: ButtonBuilder;
 
-export const customId = "declaration_second";
-export const data = new ButtonBuilder()
-  .setCustomId(customId)
-  .setStyle(ButtonStyle.Secondary)
-  .setLabel("2凸目");
-
-export async function execute(interaction: ButtonInteraction) {
-  let guild: Guild;
-  if (interaction.guild != null) {
-    guild = interaction.guild;
-  } else {
-    throw new Error("interaction.guild is null");
+  constructor() {
+    super();
+    this.attackCount = 2;
+    this.data = new ButtonBuilder()
+      .setCustomId(DeclarationSecond.customId)
+      .setStyle(ButtonStyle.Secondary)
+      .setLabel("2凸目");
   }
-  if (interaction.channel == null) {
-    throw new Error("interaction.channel is null");
-  }
-  const channel = guild.channels.cache.find((channel) => channel.id === interaction.channel?.id);
-  if (channel == null) {
-    throw new Error("チャンネル情報が取得できませんでした");
-  }
-  if (channel.parentId == null) {
-    throw new Error("親カテゴリ情報が取得できませんでした");
-  }
-  const clan = await DataSource.getRepository(Clan).findOneBy({
-    discordCategoryId: channel.parentId,
-  });
-  if (clan == null) {
-    throw new Error("クラン情報が取得できませんでした");
-  }
-  // ボス情報取得
-  const bossRepository = DataSource.getRepository(Boss);
-  const boss = await bossRepository.findOneBy({
-    clanId: clan.id,
-    discordChannelId: interaction.channel.id,
-  });
-  if (boss == null) {
-    throw new Error("ボス情報が取得できませんでした");
-  }
-
-  const today = dayjs().format();
-  const event = await DataSource.getRepository(Event)
-    .createQueryBuilder("event")
-    .where("event.fromDate <= :today", { today })
-    .andWhere("event.toDate >= :today", { today })
-    .getOne();
-  if (event == null) {
-    throw new Error("クランバトル開催情報が取得できませんでした");
-  }
-
-  // 周回数取得
-  const lapRepository = DataSource.getRepository(Lap);
-  const lap = await lapRepository.findOneBy({
-    eventId: event.id,
-    clanId: clan.id,
-  });
-  let bossLap = 0;
-  if (lap != null) {
-    switch (boss.bossid) {
-      case 1:
-        bossLap = lap.boss1Lap ?? 1;
-        break;
-      case 2:
-        bossLap = lap.boss2Lap ?? 1;
-        break;
-      case 3:
-        bossLap = lap.boss3Lap ?? 1;
-        break;
-      case 4:
-        bossLap = lap.boss4Lap ?? 1;
-        break;
-      case 5:
-        bossLap = lap.boss5Lap ?? 1;
-        break;
-      default:
-        break;
-    }
-  }
-
-  let content = "";
-  const user = await Declaration.regist(boss, interaction.user.id, bossLap, 2);
-  if (user instanceof Error) {
-    content = user.message;
-  } else {
-    content = "【" + bossLap + "週目】" + user?.name + "が" + boss.bossid + "ボスに凸宣言しました";
-  }
-
-  const declarations = await DataSource.getRepository(DeclarationRepository).find({
-    where: {
-      bossId: boss.id,
-      isFinished: false,
-    },
-    relations: {
-      user: true,
-    },
-  });
-  const eventBossRepository = DataSource.getRepository(EventBoss);
-  const eventBoss = await eventBossRepository.findOneBy({
-    clanId: clan.id,
-    eventId: event.id,
-  });
-  if (eventBoss == null) {
-    throw new Error("クランバトルボスのHP情報が取得できませんでした");
-  }
-  await BossChannelMessage.sendMessage(
-    interaction.channel,
-    clan,
-    boss,
-    eventBoss,
-    lap,
-    declarations,
-  );
-  const deleteMessage = await channel.messages.fetch(
-    interaction.message.reference?.messageId ?? "",
-  );
-  await deleteMessage.delete();
-  await interaction.reply({ content: content });
 }
-
-export default {
-  customId,
-  data,
-  execute,
-};
